requirements.txt
# ── Core bot dependencies (original) ─────────────────────────────────────────
requests>=2.31.0          # HTTP client for Binance REST API communication
python-dotenv>=1.0.0      # Secure .env credential loading

# ── AI / ML dependencies ──────────────────────────────────────────────────────
tensorflow>=2.15.0        # LSTM neural network (Keras API)
scikit-learn>=1.4.0       # MinMaxScaler, train_test_split
pandas>=2.0.0             # DataFrame for OHLCV data and feature engineering
numpy>=1.26.0             # Array operations for model input/output sequences
